requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
email-validator==2.1.0
requests==2.31.0
Werkzeug==2.3.7
# WeasyPrint es opcional en PythonAnywhere (mejor quitarlo si da problemas)
# weasyprint==60.2
icalendar==5.0.11
gunicorn
psycopg2-binary
python-dotenv
pdfkit